Add ci.ocaml.org deployment configuration
This is based on the compose file, but uses Docker stack. This should be
more reliable because Docker itself will restart the services on
failure, or on reboot of the server. It also makes it easier to update
individual services.

The new configuration adds monitoring with Prometheus and Grafana, uses
named volumes instead of local directories, and uses Docker secrets for
the GitHub token.